refactor: change folders to constants
