+++ src/pack.ts
@@ -15,7 +15,7 @@ import {
 } from 'ramda';
 import * as semver from 'semver';
 import { EsbuildServerlessPlugin } from '.';
-import { SERVERLESS_FOLDER } from './folders';
+import { SERVERLESS_FOLDER } from './constants';
 import { doSharePath, flatDep, getDepsFromBundle } from './helper';
 import * as Packagers from './packagers';
 import { IFiles } from './types';
